// frontend_temp/src/types/global.d.ts
export interface ApiResponse {
	data: unknown,
	errors?: any,
	stack?: any,
	success: boolean,
	statusCode: number,
	message?: string
}
// For use in class page

export type SectionsType = {
	_id: string;
	name: string;
}[]

export interface ClassType {
	_id: string;
	name: string;
	sections: SectionsType;
}

// for use in api
export interface ClassesArray extends ApiResponse {
	data: ClassType[];
}

export interface SingleResponseClassType extends ApiResponse {
	data: ClassType
}
// For current user api in rtk 
export interface CurrUserSession extends ApiResponse {
	data: User
}

export interface LoginSession extends ApiResponse {
	data: { user: User }
}

export interface ApiError {
	data: ApiResponse;
	status: number;
}


export interface NoDataResponseType {
	statusCode: number,
	message: string,
	success: boolean
}

export interface DialogBoxProps<InputData> {
	title?: string;
	description?: string;
	open?: boolean;
	setOpen?: (open: boolean) => !open;
	data?: InputData;
}

export interface ProgrammeType {
	_id: string;
	name: string;
}